requirements.txt
arrow==0.17.0
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.5
djongo==1.3.3
envparse==0.2.0
Flask==1.1.2
Flask-Babel==2.0.0
Flask-WTF==0.14.3
flup==1.0.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
polib==1.1.0
pycryptodome==3.9.8
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.2.4
termcolor==1.1.0
tqdm==4.50.2
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3